+++ mcpath/facades/java.py
@@ -6,6 +6,7 @@ Linux, MacOS, Windows
 
 __all__ = ["Java"]
 
+from typing import Optional
 from os import path
 
 from ..utils import deprecated
@@ -136,49 +137,49 @@ class Java:
     def _launch(self):
         return None
 
-    def _get_runtime(self, version):
+    def _get_runtime(self, version) -> Optional[str]:
         return None
 
-    def _get_root_dir(self):
+    def _get_root_dir(self) -> Optional[str]:
         return None
 
-    def _get_launcher(self):
+    def _get_launcher(self) -> Optional[str]:
         return None
 
-    def _get_game_dir(self):
+    def _get_game_dir(self) -> Optional[str]:
         return None
 
-    def _get_versions_dir(self):
+    def _get_versions_dir(self) -> Optional[str]:
         root_dir = self.get_root_dir()
         if root_dir is None:
             return None
         return path.join(root_dir, "versions")
 
-    def _get_saves_dir(self):
+    def _get_saves_dir(self) -> Optional[str]:
         game_dir = self.get_game_dir()
         if game_dir is None:
             return None
         return path.join(game_dir, "saves")
 
-    def _get_resource_packs_dir(self):
+    def _get_resource_packs_dir(self) -> Optional[str]:
         game_dir = self.get_game_dir()
         if game_dir is None:
             return None
         return path.join(game_dir, "resourcepacks")
 
-    def _get_screenshots_dir(self):
+    def _get_screenshots_dir(self) -> Optional[str]:
         game_dir = self.get_game_dir()
         if game_dir is None:
             return None
         return path.join(game_dir, "screenshots")
 
-    def _get_backups_dir(self):
+    def _get_backups_dir(self) -> Optional[str]:
         game_dir = self.get_game_dir()
         if game_dir is None:
             return None
         return path.join(game_dir, "backups")
 
-    def _get_logs_dir(self):
+    def _get_logs_dir(self) -> Optional[str]:
         game_dir = self.get_game_dir()
         if game_dir is None:
             return None