+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.2] - 10/28/2025
+
+### General
+
+- Added support for new GDK path for Bedrock and Preview for Windows
+
+### New
+
+- Added function get_edition
+- Added function has_edition
+
 ## [2.0.1] - 5/7/2025
 
 ### Fixes